fix: use count(*)::int cast per CLAUDE.md anti-patterns
Fixes both findByOwner (pre-existing) and findPendingSuggestions (new)
to use ::int cast so Postgres returns a number, not a string.

+++ src/server/repositories/transaction.repository.ts
@@ -59,7 +59,7 @@ export class TransactionRepository
     });
 
     const [{ count: total }] = await this.db
-      .select({ count: sql<number>`count(*)` })
+      .select({ count: sql<number>`count(*)::int` })
       .from(transactions)
       .where(whereClause);
 
@@ -258,7 +258,7 @@ export class TransactionRepository
     });
 
     const [{ count: total }] = await this.db
-      .select({ count: sql<number>`count(*)` })
+      .select({ count: sql<number>`count(*)::int` })
       .from(transactions)
       .where(whereClause);
